Refactor so that RAG chain is separated and encapsulated
